\name{mvama-package}
\alias{mvama-package}
\alias{mvama}
\docType{package}
\title{
Multivariate analysis of microarrays
}
\description{
Analyzes a microarray data set in the context of a gene network. Correlations between genes
are explicitly acounted for.
}
\details{
\tabular{ll}{
Package: \tab mvama\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-01-19\cr
License: \tab GPL(>=2)\cr
LazyLoad: \tab yes\cr
}
Four steps are involved in the analysis. Fitting a sparse inverse covariance matrix, constructing permutation 
distributions, performing significance tests, and plotting graphs of interesting network regions or relationships.
}
\author{
Harri Kiiveri

Maintainer:  <harri.kiiveri@csiro.au>
}
\references{
Kiiveri(2011) Multivariate analysis of microarray data: differential expression and differential connection. BMC Bioinformatics.
}
\keyword{ package }
\seealso{
}
\examples{
# see help files
}
