\name{perms}
\alias{perms}
\title{
Compute samples from the null distribution of contrasts by permutation
}
\description{
Compute samples from the null distribution of contrasts by permutation. Also compute the observed values of the contrasts.
}
\usage{
perms(C, X, D, nperm = 500, seed = 123)
}
\arguments{
  \item{C}{
matrix - r by s contrast matrix. This must be full rank.
}
  \item{X}{
matrix - n by p data matrix.
}
  \item{D}{
matrix - n by r design matrix for the experiment.
}
  \item{nperm}{
scalar - the number of permutations required.
}
  \item{seed}{
scalar - starting seed for the permutations.
}
}
\details{
Null distributions are generated by the permutation method described in Kiiveri(2011) below.
}
\value{
A list with components
\item{CTB }{a list of length s. Each component is a p by nperm matrix storing realisations of the contrast under the null hypothesis.}
\item{DQ }{a n by r matrix - the transformed design matrix.}
\item{ctb0 }{a p by s matrix - the observed values of the contrasts}
\item{nperm }{the number of permutations}
\item{seed }{the starting seed for the permutations}
}
\references{
Kiiveri. H. T.(2011) Multivariate analysis of microarray data: differential expression and differential connection. BMC Bioinformatics.
}
\author{
Harri Kiiveri
}
\note{
%%  ~~further notes~~
}
\seealso{
\code{\link{perm.summary}}
}
\examples{
# small 3 group problem with 5 variables and 150 samples
# generate mean matrix
a<-as.numeric(gl(3,50,150))
b<-c(0,1,0,1,1)
M<-outer(a,b,"*") # variables 2,4 and 5 have nonzero and unequal group means
# design matrix
D<-matrix(0,nrow=150,ncol=3)
D[1:50,1]<-1
D[51:100,2]<-1
D[101:150,3]<-1
# generate data
X<-matrix(rnorm(5*150),ncol=5)+M
# contrast matrix for testing equality of group2 and 1 and group 3 and group 1 means
C<-matrix(c(1,-1,0,1,0,-1)/2^0.5,ncol=2)
# generate permutation distributions
res<-perms(C,X,D,nperm=100,seed=123)
#check components
length(res$CTB)
dim(res$CTB[[1]])
#observed contrast values
res$ctb0
}
\keyword{ multivariate }
\keyword{distribution}
